Moving the annotation layer
