+++ client/app/reader/LoadingScreen/LoadingScreenActions.js
@@ -1,6 +1,7 @@
 import * as Constants from '../constants';
 import { collectAllTags } from '../PdfViewer/PdfViewerActions';
 import { setViewedAssignment } from '../CaseSelect/CaseSelectActions';
+import { updateFilteredIds } from '../commonActions';
 
 export const onInitialDataLoadingFail = (value = true) => ({
   type: Constants.REQUEST_INITIAL_DATA_FAILURE,
@@ -19,10 +20,6 @@ export const setLoadedVacolsId = (vacolsId) => ({
   }
 });
 
-export const updateFilteredIds = () => ({
-  type: Constants.UPDATE_FILTERED_DOC_IDS
-});
-
 export const onReceiveDocs = (documents, vacolsId) =>
   (dispatch) => {
     dispatch({
@@ -36,11 +33,6 @@ export const onReceiveDocs = (documents, vacolsId) =>
     dispatch(setViewedAssignment(vacolsId));
     dispatch(setLoadedVacolsId(vacolsId));
   };
-
-export const onReceiveAnnotations = (annotations) => ({
-  type: Constants.RECEIVE_ANNOTATIONS,
-  payload: { annotations }
-});
 
 export const onReceiveManifests = (manifestVbmsFetchedAt, manifestVvaFetchedAt) => ({
   type: Constants.RECEIVE_MANIFESTS,
