+++ client/app/reader/ReaderLoadingScreen.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { bindActionCreators } from 'redux';
 import { ENDPOINT_NAMES } from './analytics';
 import ApiUtil from '../util/ApiUtil';
-import { onReceiveDocs, onReceiveManifests, onReceiveAnnotations, onInitialDataLoadingFail
+import { onReceiveDocs, onReceiveManifests, onInitialDataLoadingFail
 } from './LoadingScreen/LoadingScreenActions';
+import { onReceiveAnnotations } from './AnnotationLayer/AnnotationActions';
 import { connect } from 'react-redux';
 import StatusMessage from '../components/StatusMessage';
 import LoadingScreen from '../components/LoadingScreen';
