+++ client/app/reader/commonActions.js
@@ -1,6 +1,7 @@
 import {
   HIDE_ERROR_MESSAGE,
-  SHOW_ERROR_MESSAGE
+  SHOW_ERROR_MESSAGE,
+  UPDATE_FILTERED_DOC_IDS
 } from './constants';
 
 // errors
@@ -16,5 +17,14 @@ export const showErrorMessage = (messageType) => ({
   type: SHOW_ERROR_MESSAGE,
   payload: {
     messageType
+  }
+});
+
+// Apply filters
+
+export const updateFilteredIds = (annotationLayer = {}) => ({
+  type: UPDATE_FILTERED_DOC_IDS,
+  payload: {
+    annotationLayer
   }
 });
